game_of_life source code and solution
