+++ solutions/game_of_life/game_of_life.py
